ROFO-31 refactor: 위치 권한 datasource로 분리
