+++ data/src/main/java/com/weit2nd/data/di/PositionModule.kt
@@ -3,6 +3,7 @@ package com.weit2nd.data.di
 import android.content.Context
 import com.google.android.gms.location.LocationServices
 import com.weit2nd.data.repository.position.CurrentPositionRepositoryImpl
+import com.weit2nd.data.source.position.LocationPermissionDataSource
 import com.weit2nd.domain.repository.position.CurrentPositionRepository
 import dagger.Module
 import dagger.Provides
@@ -19,9 +20,17 @@ object PositionModule {
     @Provides
     fun provideCurrentPositionRepository(
         @ApplicationContext context: Context,
+        locationPermissionDataSource: LocationPermissionDataSource,
     ): CurrentPositionRepository {
         return CurrentPositionRepositoryImpl(
+            locationPermissionDataSource = locationPermissionDataSource,
             locationClient = LocationServices.getFusedLocationProviderClient(context),
         )
+    }
+
+    @ViewModelScoped
+    @Provides
+    fun providesLocationPermissionDataSource(): LocationPermissionDataSource {
+        return LocationPermissionDataSource()
     }
 }
