+++ data/src/main/java/com/weit2nd/data/repository/position/CurrentPositionRepositoryImpl.kt
@@ -1,6 +1,5 @@
 package com.weit2nd.data.repository.position
 
-import android.Manifest
 import android.annotation.SuppressLint
 import android.os.Looper
 import com.google.android.gms.location.FusedLocationProviderClient
@@ -8,7 +7,7 @@ import com.google.android.gms.location.LocationCallback
 import com.google.android.gms.location.LocationRequest
 import com.google.android.gms.location.LocationResult
 import com.google.android.gms.location.Priority
-import com.gun0912.tedpermission.coroutine.TedPermission
+import com.weit2nd.data.source.position.LocationPermissionDataSource
 import com.weit2nd.domain.model.Location
 import com.weit2nd.domain.repository.position.CurrentPositionRepository
 import kotlinx.coroutines.CancellableContinuation
@@ -17,22 +16,14 @@ import javax.inject.Inject
 import kotlin.coroutines.resume
 
 class CurrentPositionRepositoryImpl @Inject constructor(
+    private val locationPermissionDataSource: LocationPermissionDataSource,
     private val locationClient: FusedLocationProviderClient,
 ) : CurrentPositionRepository {
-
-    private val permission = TedPermission.create()
-        .setDeniedTitle("위치 권한 거절")
-        .setDeniedMessage("권한 설정을 해주셔야 현재 위치 파악이 가능합니다.")
-        .setGotoSettingButtonText("설정")
-        .setPermissions(
-            Manifest.permission.ACCESS_FINE_LOCATION,
-            Manifest.permission.ACCESS_COARSE_LOCATION
-        )
 
     override suspend fun getCurrentPosition(): Location {
         val defaultLocation = Location(latitude = 37.5597706, longitude = 126.9423666)
 
-        if (permission.checkGranted()) {
+        if (locationPermissionDataSource.requestLocationPermission()) {
             return suspendCancellableCoroutine { continuation ->
                 requestLocationUpdates(continuation)
             }
